pyproject.toml
[tool.poetry]
name = "easy-crud-repo-service"
version = "0.1.1"
description = ""
authors = ["Smolke <d.smolczynski1@gmail.com>"]
readme = "README.md"
packages = [{include = "easy_crud_repo_service"}]

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^8.0.33"
python-dotenv = "^1.0.0"
inflection = "^0.5.1"
easyvalid-data-validator = "^2.0.1"
dbm-database-service = "^0.2.5"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
coverage = "^7.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
